+++ libnopegl/src/gen_specs.c
@@ -314,10 +314,10 @@ static int check_node_params(const struct node_class *cls)
         return NGL_ERROR_BUG;
     }
 
-    if (cls->category == NGLI_NODE_CATEGORY_TRANSFORM) {
+    if (cls->category == NGLI_NODE_CATEGORY_TRANSFORM && cls->draw) {
         if (strcmp(par->key, "child") ||
             !(par->flags & NGLI_PARAM_FLAG_NON_NULL)) {
-            fprintf(stderr, "transform nodes are expected to have a non-null child as first parameter");
+            fprintf(stderr, "drawable transform nodes are expected to have a non-null child as first parameter");
             return NGL_ERROR_BUG;
         }
     }