+++ libnopegl/src/transforms.h
@@ -24,18 +24,6 @@
 
 #include "internal.h"
 
-#define TRANSFORM_TYPES_ARGS                    NGL_NODE_ROTATE,    \
-                                                NGL_NODE_ROTATEQUAT,\
-                                                NGL_NODE_TRANSFORM, \
-                                                NGL_NODE_TRANSLATE, \
-                                                NGL_NODE_SCALE,     \
-                                                NGL_NODE_SKEW
-
-#define TRANSFORM_TYPES_LIST (const uint32_t[]){TRANSFORM_TYPES_ARGS, NGL_NODE_IDENTITY, NGLI_NODE_NONE}
-
-const struct ngl_node *ngli_transform_get_leaf_node(const struct ngl_node *node);
-int ngli_transform_chain_check(const struct ngl_node *node);
-void ngli_transform_chain_compute(const struct ngl_node *node, float *matrix);
 void ngli_transform_draw(struct ngl_node *node);
 
 #endif