+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Versioning](https://semver.org/spec/v2.0.0.html) for `libnopegl`.
 - Android binding and viewer
 - Automatic color conversions from one set of primaries to another
 - DrawRect node
+- AffineTransform node
 
 ### Fixed
 - Crash when using resizable RTTs with time ranges
@@ -83,6 +84,7 @@ Versioning](https://semver.org/spec/v2.0.0.html) for `libnopegl`.
   `nopegl/nopegl_android.h`, `nopegl/nopegl_opengl.h`
 - Android-specific APIs are now only available on Android
 - `scene_params.aspect_ratio` -> `scene_params.{width,height}`
+- Texture reframing is now done through a separate parameter
 
 ### Removed
 - `Text.aspect_ratio` and `DrawPath.aspect_ratio`, they now match the scene
@@ -90,6 +92,7 @@ Versioning](https://semver.org/spec/v2.0.0.html) for `libnopegl`.
 - `GraphicConfig.scissor_test` parameter
 - Support for Android < 9.0
 - The `small` build option (enabled by default now)
+- Chained transforms
 
 ## [2024.0 / libnopegl 0.11.0][2024.0] - 2024-02-02
 ### Added